requirements.txt
pandas>=1.5.0
numpy>=1.21.0
fsspec>=2023.1.0
adlfs>=2023.1.0
seaborn>=0.11.0
matplotlib>=3.5.0
openai>=1.0.0
azure-identity>=1.12.0
azure-ai-ml>=1.8.0
azureml-core>=1.50.0
azure-storage-blob>=12.0.0
aiofiles>=23.0.0
scikit-learn>=1.0.0